Show coach number and track number
If an outbound train has a track number, we show that above the "departs
in" message. If an outbound train doesn't have a track number yet, but
has a coach number, we show that instead.

We only show this information for outbound trains, since it's really
only relevant when you're departing from a station with multiple tracks
(North or South Station).