Fix bug showing "departing now" for 1h 0m
Fixes https://github.com/thoughtbot/PurpleTrainElm/issues/14

This removes the elm-date-extra library, since it has a buggy
implementation of `Duration.diff`. The documentation for elm-date-extra
claims that this function returns a delta record with either all
positive values or all negative values, but we found that not to be the
case, which caused our logic to be wrong for showing the relative time
message.

This also adds an `Api` module to make it easier to point the
application to a separate base URL, e.g. to test the app against a local
API server.