Autoscroll to the currently selected Stop
The StopView always started at the top. This was confusing and inconvenient
for people who need to pick stops close to each other and/or people who
accidentally hit the wrong stop.

This change makes it so that the currentlu selected stop is scrolled into the
center of the StopView box when its opened. If the selected stop is too near
the top or bottom of the list, the list is not scrolled past the limits you
would expect a list of this type to scroll past.

The stop is also highlighted in the color #F5DEFF which is a light purple.
This might not be the ideal color, but it's visible highlighted. We can change
it if we need to. Honestly, it doesn't look great, but them's the breaks when
you have a developer do design work and knows that it can easily be changed
later. It gets done just enough to prove the concept. So there.