Add sync script for Elm Native UI
Since many of us are working on the Elm Native UI (ENUI) library, this
provides an easier way to manage the dependency than the alternatives.

The elm_self_publish script from
https://github.com/NoRedInk/elm-ops-tooling creates a submodule, which
is a pain to manage, doesn't get automatically updated, and causes big
diffs if you switch out the directory to work on ENUI.

We'd previously deleted the submodule and kept the entire directory
checked in, but this faces similar problems of causing big git diffs
when working on the library itself.

This solution removes the dependency from this repo entirely, and uses a
sync script to keep things up to date. This allows you to keep a
symlinked directory if you are working on ENUI. If you _don't_ have a
symlinked directory, then running the script will either clone or update
the dependency to the version specified in `.elm-native-ui-version`.