Hack to show border radius on Android
For some reason the bottom of the stops list renders with ugly square
corners on Android instead of the specified border radius from the
picker container.

This hack adds a 10px empty view below the list view for Android, with
border radius set.

Hopefully we can remove this at some point in favor of a better
solution.