Add GPL license
The intent is to open source this application to serve as a showcase for
our Elm work and as a general example of a real-world elm-native-ui application.
We want to avoid a situation where someone else takes the source code and sells
it without making the source available.

Also, add info about thoughtbot to README