Display arrival times
When a train leaves the station it has to stop somewhere. We should assume
it's going to stop where you're intending to go. On the Elixir side, we're
including the arrival time of the train (as per
thoughtbot/purple_train_web#48). This commit displays this arrival time just
below the departure time. It currently messes up the layout a bit but that can
be modified.

This must be applied AFTER
https://github.com/thoughtbot/purple_train_web/pull/48 or we'll get errors.