Support iOS push notifications for cancellations
This prompts the user to give permission for push notifications so they
can be notified when a train is cancelled that would've gone to/from
their home stop.

Once the user accepts the permission dialog, we get their device token.
Once they select their home stop, we send a request to our API to upsert
an "installation", which is what the backend uses to know which devices
to send push notifications to, and when a cancellation applies to a
given device (based on the selected stop).

The PushNotificationIOS library is not yet in elm-native-ui, so for now
we register for push notifications and get the device token in JS, then
communicate the device token to the Elm app with a port.

Here's the flow:

If we've never prompted the user to enable push notifications:
- Pre-prompt them, describing why they should allow notifications
- If they accept, show actual iOS notifications prompt
- If they accept that:
  - If they already have a stop selected (existing PT user), upsert
    their installation via the API so the API can send them
    notifications
  - Send deviceToken to the elm app via a port so the Elm app can upsert
    the installation when a stop is selected
  - Store deviceToken in AsyncStorage so the app can load it as a
    setting upon launch